Separating out the virtualized table component
